+++ src/main.py
@@ -1,6 +1,5 @@
 
-from FunFiles.file_reader import FileReader
-from csv_reader import CsvReader
+from file_reader import FileReader
 from text_reader import TextReader
 
 
@@ -10,8 +9,6 @@ def main():
     path.name()
     print_file = TextReader(r"C:\Users\User\Documents\meow.txt")
     # print_file.read()
-    csv_file = CsvReader(r"C:\Users\User\Documents\csv_file.csv")
-    csv_file.read(6)
 
 
 if __name__ == '__main__':
